RoadSignNet-SAL-main/requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
albumentations>=1.3.1

# Training & Monitoring
tensorboard>=2.13.0
tqdm>=4.65.0
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Dataset Management
roboflow>=1.1.0

# Utilities
scikit-learn>=1.3.0
scipy>=1.11.0